apps/worker/requirements.txt
celery
redis
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
httpx
openai
